pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rivet-gs"
version = "0.1.0"
description = "RIVET Gene Study (HPO mapping) Sejong University Academic Festival Edition"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name="JIWOOK HA" }]
dependencies = ["pandas>=1.3", "pyyaml>=6.0"]

[project.scripts]
rivet-gs = "rivet_gs.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rivet_gs*"]
